Fixing discounts with higher count items
This change allows the discounter to work with items when the count
is higher than the requirement.

Previously it would only discount itmes that were exactly what the
requirements were (eg. two_for_one requires a count of 2). Now it
will apply the discount and add the values of the other non-discounted
items.